WIP: start to address review
