+++ frontend/stylesheets/main.less
@@ -110,18 +110,6 @@ body {
   }
 }
 
-.ant-alert-message {
-  margin-left: 5px;
-}
-.ant-alert-banner {
-  position: absolute;
-  z-index: 1000;
-  width: 100%;
-}
-.ant-dropdown-menu-item-icon.fa-icon {
-  margin-right: 5px;
-}
-
 .fa-random {
   width: 14px;
   background-position: -217px -120px;
